# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tienda Magic API",
        "description": "Documentación de la API de Tienda Magic Swagger OpenApi",
        "contact": {
            "email": "soporte@tiendamagic.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exchanges": {
            "get": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Listar intercambios",
                "description": "Obtiene los intercambios activos en el mercado.",
                "operationId": "0d0e2fddd1de7add3491f41b3c71144a",
                "responses": {
                    "200": {
                        "description": "Lista de intercambios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Crear listado de intercambio",
                "description": "El usuario ofrece una carta de su inventario pidiendo opcionalmente una carta específica a cambio.",
                "operationId": "e5a92eb4d4fe4d7ee1b58ff2ad54794f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "offered_inventory_card_id": {
                                        "type": "integer"
                                    },
                                    "requested_card_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Listado de intercambio creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exchanges/{id}/request": {
            "post": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Solicitar un intercambio",
                "description": "Envía una solicitud proponiendo una carta a un listado de intercambio activo.",
                "operationId": "64fe871e1c56cc90ea6648206f593478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del intercambio activo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "offered_inventory_card_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de intercambio creada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/forums/{forum}": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Foro y sus threads",
                "description": "Muestra los threads de un foro concreto, paginados y ordenados por parámetro (hot, new, top).",
                "operationId": "9e76ecaff092f4714088760bb399e134",
                "parameters": [
                    {
                        "name": "forum",
                        "in": "path",
                        "description": "ID del foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tipo de ordenamiento (hot, new, top)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "hot"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Threads del foro"
                    }
                }
            }
        },
        "/api/saved": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Lista de hilos guardados",
                "description": "Devuelve los hilos que el usuario ha guardado.",
                "operationId": "d282c4bd41591796d22beb72071f6101",
                "responses": {
                    "200": {
                        "description": "Lista paginada de hilos guardados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/saved/{thread}": {
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Guardar hilo",
                "description": "Añade un hilo a la lista de guardados del usuario.",
                "operationId": "b54b9de5fe4c767a3ce4c43343e75cc4",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo guardado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Forums"
                ],
                "summary": "Eliminar hilo guardado",
                "description": "Saca un hilo de la lista de guardados del usuario.",
                "operationId": "43e75768a553b2a22ef934bd0b0eb538",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo eliminado exitosamente de guardados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}